fix deleting log file even if it doesn't exist

+++ Sibyl_System/__init__.py
@@ -11,8 +11,8 @@ import logging
 import os
 import re
 
-
-os.remove("log.txt")
+if os.path.exists('log.txt')
+    os.remove("log.txt")
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
